# back-end/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Education website with Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "name@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/present/createPresent": {
            "post": {
                "description": "Return present's informaion.",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Present category's id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Present's Name",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Present's Image",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Present's Price",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Present's Score",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Present's Description",
                        "required": true,
                        "type": "text"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create present successfully",
                        "schema": {
                            "properties": {
                                "category": {
                                    "type": "integer"
                                },
                                "image": {
                                    "type": "text"
                                },
                                "name": {
                                    "type": "text"
                                },
                                "price": {
                                    "type": "text"
                                },
                                "author": {
                                    "type": "integer"
                                },
                                "description": {
                                    "type": "text"
                                },
                                "score": {
                                    "type": "text"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    },
                    "401": {
                        "description": "You are not admin!!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/present/getAllPresent": {
            "get": {
                "description": "Return all present's informaion.",
                "responses": {
                    "200": {
                        "description": "Get present successfully",
                        "schema": {
                            "properties": {
                                "category": {
                                    "type": "integer"
                                },
                                "image": {
                                    "type": "text"
                                },
                                "name": {
                                    "type": "text"
                                },
                                "price": {
                                    "type": "text"
                                },
                                "author": {
                                    "type": "integer"
                                },
                                "description": {
                                    "type": "text"
                                },
                                "score": {
                                    "type": "text"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/present/getOnePresent/id": {
            "get": {
                "description": "Return one present's informaion.",
                "responses": {
                    "200": {
                        "description": "Get present successfully",
                        "schema": {
                            "properties": {
                                "category": {
                                    "type": "integer"
                                },
                                "image": {
                                    "type": "text"
                                },
                                "name": {
                                    "type": "text"
                                },
                                "price": {
                                    "type": "text"
                                },
                                "author": {
                                    "type": "integer"
                                },
                                "description": {
                                    "type": "text"
                                },
                                "score": {
                                    "type": "text"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id!!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/present/updatePresent/{id}": {
            "post": {
                "description": "Return present's informaion.",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Present category's id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Present's Name",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Present's Image",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Present's Price",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Present's Score",
                        "required": true,
                        "type": "text"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Present's Description",
                        "required": true,
                        "type": "text"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update present successfully",
                        "schema": {
                            "properties": {
                                "category": {
                                    "type": "integer"
                                },
                                "image": {
                                    "type": "text"
                                },
                                "name": {
                                    "type": "text"
                                },
                                "price": {
                                    "type": "text"
                                },
                                "author": {
                                    "type": "integer"
                                },
                                "description": {
                                    "type": "text"
                                },
                                "score": {
                                    "type": "text"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id !!!"
                    },
                    "401": {
                        "description": "You are not admin !!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}